\name{qPCRraw}
\Rdversion{1.1}
\docType{data}
\alias{qPCRraw}
\title{Example raw qPCR data.}
\description{Six qPCR samples, performed on the TaqMan Low Density Arrays from Applied Biosystem. Each sample contains 384 PCR reactions, and there are 3 different samples with 2 replicates each. To be used as example data in the HTqPCR package.}

\usage{data(qPCRraw)}

\format{
  An object of class qPCRset. The format is:
Formal class 'qPCRset' [package ".GlobalEnv"] with 9 slots
  ..@ featureNames   : chr [1:384] "Gene1" "Gene2" "Gene3" "Gene4" ...
  ..@ sampleNames    : chr [1:6] "sample1" "sample2" "sample3" "sample4" ...
  ..@ exprs          : num [1:384, 1:6] 11.5 33.9 28 26.9 25 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:384] "Gene1" "Gene2" "Gene3" "Gene4" ...
  .. .. ..$ : chr [1:6] "sample1" "sample2" "sample3" "sample4" ...
  ..@ flag           :'data.frame':	384 obs. of  6 variables:
  .. ..$ V1: chr [1:384] "Passed" "Passed" "Passed" "Passed" ...
  .. ..$ V2: chr [1:384] "Passed" "Passed" "Passed" "Passed" ...
  .. ..$ V3: chr [1:384] "Passed" "Passed" "Passed" "Passed" ...
  .. ..$ V4: chr [1:384] "Flagged" "Flagged" "Passed" "Passed" ...
  .. ..$ V5: chr [1:384] "Passed" "Passed" "Passed" "Passed" ...
  .. ..$ V6: chr [1:384] "Passed" "Passed" "Passed" "Passed" ...
  ..@ featureType    : Factor w/ 2 levels "Endogenous Control",..: 1 2 2 2 2 2 2 2 2 2 ...
  ..@ featurePos     : chr [1:384] "A1" "A2" "A3" "A4" ...
  ..@ featureClass   : Factor w/ 3 levels "Kinase","Marker",..: 3 3 2 1 2 3 1 3 3 3 ...
  ..@ featureCategory:'data.frame':	384 obs. of  6 variables:
  .. ..$ X1: chr [1:384] "OK" "OK" "OK" "OK" ...
  .. ..$ X2: chr [1:384] "OK" "OK" "OK" "OK" ...
  .. ..$ X3: chr [1:384] "OK" "OK" "OK" "OK" ...
  .. ..$ X4: chr [1:384] "OK" "OK" "OK" "OK" ...
  .. ..$ X5: chr [1:384] "OK" "OK" "OK" "OK" ...
  .. ..$ X6: chr [1:384] "OK" "OK" "OK" "OK" ...
  ..@ history        :'data.frame':	1 obs. of  1 variable:
  .. ..$ history: chr "Default HTqPCR qPCRset object with raw data."
}

\examples{
data(qPCRraw)
}

\keyword{datasets}
