\name{qPCRpros}
\Rdversion{1.1}
\alias{qPCRpros}
\docType{data}

\title{Example processed qPCR data}
\description{Processed version of the raw data in qPCRraw, to be used as example data in the HTqPCR package. The data has been processed with \code{setCategory} to mark the feature categories, and with \code{normalizaHTqPCRCard} using rank invariant normalisation.}

\usage{data(qPCRpros)}

\format{
  The format is:
Formal class 'qPCRset' [package ".GlobalEnv"] with 9 slots
  ..@ featureNames   : chr [1:384] "Gene1" "Gene2" "Gene3" "Gene4" ...
  ..@ sampleNames    : chr [1:6] "sample1" "sample2" "sample3" "sample4" ...
  ..@ exprs          : num [1:384, 1:6] 11.5 33.9 28 26.9 25 ...
  .. ..- attr(*, "dimnames")=List of 2
  .. .. ..$ : chr [1:384] "Gene1" "Gene2" "Gene3" "Gene4" ...
  .. .. ..$ : chr [1:6] "sample1" "sample2" "sample3" "sample4" ...
  ..@ flag           :'data.frame':	384 obs. of  6 variables:
  .. ..$ V1: chr [1:384] "Passed" "Passed" "Passed" "Passed" ...
  .. ..$ V2: chr [1:384] "Passed" "Passed" "Passed" "Passed" ...
  .. ..$ V3: chr [1:384] "Passed" "Passed" "Passed" "Passed" ...
  .. ..$ V4: chr [1:384] "Flagged" "Flagged" "Passed" "Passed" ...
  .. ..$ V5: chr [1:384] "Passed" "Passed" "Passed" "Passed" ...
  .. ..$ V6: chr [1:384] "Passed" "Passed" "Passed" "Passed" ...
  ..@ featureType    : Factor w/ 2 levels "Endogenous Control",..: 1 2 2 2 2 2 2 2 2 2 ...
  ..@ featurePos     : chr [1:384] "A1" "A2" "A3" "A4" ...
  ..@ featureClass   : Factor w/ 3 levels "Kinase","Marker",..: 3 3 2 1 2 3 1 3 3 3 ...
  ..@ featureCategory:'data.frame':	384 obs. of  6 variables:
  .. ..$ X1: chr [1:384] "Unreliable" "OK" "OK" "OK" ...
  .. ..$ X2: chr [1:384] "Unreliable" "Undetermined" "OK" "OK" ...
  .. ..$ X3: chr [1:384] "Unreliable" "OK" "OK" "OK" ...
  .. ..$ X4: chr [1:384] "OK" "OK" "OK" "OK" ...
  .. ..$ X5: chr [1:384] "Unreliable" "Undetermined" "OK" "OK" ...
  .. ..$ X6: chr [1:384] "OK" "OK" "OK" "OK" ...
  ..@ normalized     : chr "rankinvariant"
}

\examples{
data(qPCRpros)
}

\keyword{datasets}
